chore(advanced_chunks): move related code to a seperate file
